# OpenApiStandards/ConsoleAppClient/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPIWeb Demo",
    "description": "Application to showcase OpenAPI Spec Please consider using later version for new development.",
    "termsOfService": "https://www.keefegroup.com/home/terms-conditions-108",
    "contact": {
      "name": "COL Team",
      "url": "https://www.keefegroup.com/home/contact-us-105",
      "email": "CommissaryOnline@keefegroup.com"
    },
    "license": {
      "name": "License Keefe Group",
      "url": "https://www.keefegroup.com/"
    },
    "version": "1.0"
  },
  "paths": {
...
    "/WeatherForecast/{city}": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "summary": "Gets forecast for all the cities",
        "operationId": "GetByCity",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "description": "City where you want the forecast for",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the weather for the city",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OpenApiWeb.WeatherForecast"
                  }
                }
              }
            }
          },
          "404": {
            "description": "If the city not matched",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/WeatherForecast/{city}/{zip}": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetByCityZip",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "zip",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/WeatherForecast/{city}/extended": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetByCityExtended",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenApiWeb.WeatherForecastExtended"
                }
              }
            }
          }
        }
      }
    },
    "/WeatherForecast/{city}/provider/{provider}": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetByCityAndProvider",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenApiWeb.WeatherForecast"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Microsoft.AspNetCore.Mvc.ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "OpenApiWeb.Forecast": {
        "enum": [
          "RANDOM",
          "PROPABLE",
          "ACCURATE"
        ],
        "type": "string",
        "description": "Accuracy of weather forecast prediction"
      },
...
      "OpenApiWeb.WeatherForecast": {
        "required": [
          "city"
        ],
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "Weather Forecast Date",
            "format": "date",
            "example": "12/12/2021"
          },
          "temperatureC": {
            "type": "integer",
            "description": "Temperature in Celcius",
            "format": "int32",
            "example": 30
          },
          "temperatureF": {
            "type": "integer",
            "description": "Temperature in Fahrenheit",
            "format": "int32",
            "readOnly": true,
            "example": -1
          },
          "summary": {
            "type": "string",
            "description": "Description of weather forecast",
            "nullable": true,
            "example": "Mostly sunny"
          },
          "city": {
            "maxLength": 30,
            "minLength": 0,
            "type": "string",
            "description": "The name of the city",
            "example": "St. Louis"
          },
          "forecastAccuracy": {
            "$ref": "#/components/schemas/OpenApiWeb.Forecast"
          }
        },
        "additionalProperties": false
      },
      "OpenApiWeb.WeatherForecastExtended": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/OpenApiWeb.WeatherForecast"
          }
        ],
        "properties": {
          "next7Days": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}